The in-progress anxiety svg file
I will be adding more to this long image, but I want to see what my
stakeholder thinks of it first. It is really more aesthetic than
anything, but reinforces the concept of the content. I will need to
introduce a break point so that it is not seen on a mobile device
(iPad/Nexus okay, iPhone et al, definitely not okay!).