Lots of addition to content as a result of formatting
I am formatting the different <sections> of the page so have added
quite a bit of code to this file over the past few days. Specifically I
have fixed the smooth scroll to account for the header height. I have
developed my grid for difference sections of the page. I have developed
the Team area and have added “About this project”. I have fleshed out
placeholder content in the Flickity Carousel (this script is ignored in
the git). I will be continuing work on the sections today and also
changing the header navigation as it is not sufficiently accessible as
a fixed header.