A lot of grid specific styling is added
There will be a lot more tweaking to do in terms of keeping this
responsive. Breakpoints through media queries still require a serious
looking at. The background anxiety svg image has been added to the CSS
and I need to complete a more thorough plan of the <sections> content
and the corresponding “areas” in the CSS in order to ensure that the
baseline content is set for the breakpoints. I will be presenting the
first version of this site to the stakeholder today.