Got rid of the poll vote function
I had managed a limited amount of success with the JSON file but I
couldn’t pull the data back from it. My decision that the poll was
feature creep was by the functional issues, but instead through
reflection - I realised that it was superfluous to the needs of the
site. See journal post - on the importance of self-discipline.