New files added to support the contact form
These are php files and graphical files that Marcus provided and I have
adapted to work with my contact form. I altered my form a bit, so that
it no longer asked for first name and last name. It now only asks for
email, the reason for contact and the message text. This seemed like it
would be more streamlined for users and also when I added the FirstName
and LastName values, it wasn’t clear how I was using this information
and why it was relevant to contact. There are email addresses out there
which do not provide the user’s name, however due to the target
audience of this site folks whom we want to be in touch with will
hopefully be writing with a professional email address and include
enough information in the body of the contact form so that we can get
back to them.