Lots of changes to this version of the CSS
Streamlined CSS classes and wrote more description in about the grid
and its values. Completed more work on the media queries in terms of
moving parts of the interface when the contents is viewed on small
smartphone interfaces. I had to change the container called
flickity-page-dots to be called “.hidden” this may seem like an odd
decision, but there is method to the madness here. Through testing I
found that the Firefox clean reader mode, which uses its own local
browser-based css file, displayed the dot numbers in a problematic way
as a long list of empty numbers. As someone who uses this mode, I
wanted to fix it and also I felt that it would cause issues with
screenreaders as well. I reviewed the Firefox local CSS file and found
that there were CSS classes that could be left out of the generated
page. I had hoped to add this to the existing container class, but alas
as the dots are generated in the Javascript this was no easy task and
the hybrid names generated “dots-hidden” for example were not skipped.
The best option was to make sure that the container was completely
hidden and “hidden” was one of the exception names in the Firefox CSS
file.