Add the chart and associated styles
I’m a bit behind on my commits, so this represents work that I did with
Google Charts api a couple of weeks back. I wrote a journal post about
this, but I want to say that I really enjoyed working with the api. The
only downside is that it is quite hard to make the resulting data
visualisation as responsive as you might like. It took a long time to
get it resizing the way that I needed it to.