Worked on the media queries and added the tab
I added the tab on the User Experience (formerly scenarios) section of
the site. However, I will have to replace it because the CSS that I
used (skew) is not backwards compatible enough and it was the result of
a bit of frustrated solution on my part. It really needs to be a custom
built svg