Added more content and callout
I needed to add callout style the question and add some variety to the
text. Also I noticed a typing error in the svg that I was using for the
“return-top-button” which I thought might be affecting its display in
Internet Explorer 11. Alas it was just Internet Explorer. I did change
the dimensions later on which made this a little bit better but not
perfect in IE.