Mainly added typography styles
The typography is deliberately simple for this site, using web safe
common fonts, such as Helvetica and Arial. It gets slightly more exotic
with h4 and h5 which start with Avenir and will fallback on Arial and
Helvetica. Helvetica Neue is the main header and body font, where it is
available. Looking at medical websites, these sans-serif fonts
represent the norm and will be viewable on older less versatile web
browsers.