Refined and added a lot of written content
These additions were again about a week and a half ago. I wrote a lot
of new information for the site in anticipation of meeting with my
stakeholder. I also refined it a bit as I found that I had written too
much content! This happens to me too often, I really should cool it a
bit!