Lots of formatting changes
This was what I came to think of as the weekend of horror. I was really
struggling with some issues around alignment of the content and I found
that I needed to use some decimal values so the percentages were not
working well. They work well in the grid so I shall have to decide how
to tackle this, all “em” or all percentages. I am editing toward all
“em” for everything but the grid elements as they are working really
well with the predefined values where needed. We’ll see, I need to test
what breaks when I make these changes.