The weekend of horror
I met with my stakeholder and realised that I needed more space in the
Flickity Carousel in order to accommodate an important change to the
cancellation/rescheduling slide cell. This was an infuriating task as
the smartly produced javascript dots generate before anything else so
when the browser window was resized they would show up on top of all
the content. It was horrible. I ended up, after trying ever JavaScript
solution I could think up and lay my hands on changing the height of
the about the section margin to cover up the timeline which was firing
up before if should. It was perfect when the screen area became
smaller, but resizing the window to be bigger was not pretty before I
did this.