Got rid of some extras and wrote some new script
I got rid of the timeline height adjustment js as it was no longer
needed after the design changes to the Flickity carousel. I added some
new script for the return-top-button. This was especially important in
the wake of the new larger carousel as the appearance of the “top”
button when you were basically at the top looking at the carousel was
redundant. This script added an eventListener to watch for when the
user scrolls to more than 500 on the scrollY axis. This adds a CSS
class which shows the active button. I added an “else if” to remove the
active class in the case where scrollY is less than 500. I took out the
LazyLoader utility in the Flickity code. I tried it out first, but
found that it added effects to the images that I didn’t need and was
just taking up extra code space and memory/cache. As I need this site
to be as speedy as possible, I decided to remove it.