The placeholder anxiety svg is gone
I had been using an svg for the patient’s anxiety line which was shown
behind the content in the flickity carousel. This svg was not finished
so I had not full optimised it. However, when testing in Firefox I
discovered that the carousel’s performance was being hampered badly
(possibly more noticeable to me more than anyone else). I started
digging and using the Mozilla Dev tools to see if I could find the
problem. Low behold it was the nested div holding the super-long svg.
Also when I did optimise this svg, because it was made up of many
orphaned lines with curved ends I found that the optimiser cut the ends
up, making it look more like a bar chart! I realised that this was
probably affecting the performance, so after some experimentation
decided to draw a different type of anxiety visualisation in Adobe
Illustrator. The main goal was to use fewer lines.