Added accordion element and started to tidy and standardise
I will write about the addition of the accordion areas in more detail
in full journal post. The main gist is that I really wanted to display
the information about the devices more efficiently. I found the
accordion building tutorial “CSS based on Accordion with CSS3” by Mary
Lou. Source:
https://tympanus.net/codrops/2012/02/21/accordion-with-css3 . Although
I would normally be willing to write something like this myself (or try
anyway) it is rather late in the day and this tutorial completely fit
the bill with custom styling applied. Before I do too much on styling I
will test it on my dev PC to make sure that it doesn’t break.