This is a huge svg file and there was no way around it!
It took a better part of a day to build and customise this svg which is
the new occupant within the background of the Flickity carousel. I had
to get rid of the div with the background image, because Firefox,
Safari and IE would not register/show the image and it just resulted in
latency. The other issue is that the style of the image itself, due to
its bars layout, was creating an optical illustration which made the
latency look worse than it actually was. I ended up adding an image
source with a super long svg (this file). I found that I couldn’t
optimise this svg (without errors) as I needed to include some shadow
effects, but the good news is that the carousel is now working well.